% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage_runit.R
\name{file_coverage_runit}
\alias{file_coverage_runit}
\title{Extends the covr::file_coverage function to also accept test files based on RUnit}
\usage{
file_coverage_runit(source_files, test_files, package_name, ...)
}
\arguments{
\item{source_files}{Character vector of source files with function definitions to measure coverage}

\item{test_files}{Character vector of test files with code to test the functions}

\item{package_name}{Character with the name of the package the \code{source_files} and \code{test_files} come from}

\item{...}{Additional arguments passed to \code{\link[covr]{package_coverage}}}
}
\value{
asdf
}
\description{
Extends the covr::file_coverage function to also accept test files based on RUnit
}
