% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_coverage.R
\name{zero_coverage}
\alias{zero_coverage}
\title{Extends [covr::zero_package()] so that it also works with [filecovrunit::file_coverage_runit] outputs}
\usage{
zero_coverage(x, ...)
}
\arguments{
\item{x}{a coverage object returned [package_coverage()]}

\item{...}{additional arguments passed to
[tally_coverage()]}
}
\value{
A `data.frame` with coverage data where the coverage is 0.
}
\description{
Almost identical to [covr::zero_package()] with the only difference that it also works if your
working directory is not set to the directory where the temporary package created with [filecovrunit::file_coverage_runit]
is located.
}
\details{
if used within RStudio this function outputs the results using the
Marker API.
}
