% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage_runit.R
\name{file_coverage_runit}
\alias{file_coverage_runit}
\title{Extends the [covr::file_coverage()] function to also accept test files based on RUnit}
\usage{
file_coverage_runit(source_files, test_files, package_name, ...)
}
\arguments{
\item{source_files}{Character vector of pathes to source files with function definitions to measure coverage}

\item{test_files}{Character vector of pathes to test files with code to test the functions}

\item{package_name}{Character with the name of the package the \code{source_files} and \code{test_files} come from}

\item{...}{Additional arguments passed to [covr::package_coverage()]}
}
\value{
an covr object (identital to [covr::package_coverage()])
}
\description{
So far only works if the test-functions from the \code{test_files} are explicitally called within the same file.
}
\examples{
# load test data
src <- c(system.file(file.path("inst", "extdata", "functions.R"), package = "filecovrunit"),
system.file(file.path("inst", "extdata", "functions2.R"), package = "filecovrunit"))
test <- c(system.file(file.path("inst", "extdata", "test_1.R"), package = "filecovrunit"),
system.file(file.path("inst", "extdata", "test_2.R"), package = "filecovrunit"))
pkg_name <- "testpackage3"

# check file coverage of the files
cov <- filecovrunit::file_coverage_runit(source_files = src,
test_files = test,
package_name = pkg_name)
cov
covr::report(cov)
}
